+++ lib/src/connectors.dart
@@ -12,20 +12,45 @@ mixin WebSocketConnector on LifeCycle, Parser implements Connector {
   WebSocketChannel _socket;
   StreamSubscription _listener;
 
+  StreamController _controller;
+
+  Stream get statusStream => _controller?.stream;
+
+  @override
+  void init() {
+    _controller = StreamController.broadcast();
+    super.init();
+  }
+
   @override
   Future<void> connect(String host, int port, {dynamic args}) async {
     final useSSL = args == true;
     _socket = WebSocketChannel.connect(
       Uri(host: host, port: port, scheme: useSSL ? 'wss' : 'ws'),
     );
-    _listener = _socket.stream.listen(handleMessage);
+    _listener = _socket.stream.listen(
+      handleMessage,
+      onError: (e) {
+        _controller.addError(e);
+        disconnect();
+      },
+      cancelOnError: true,
+    );
+    _controller.add(true);
+  }
+
+  @override
+  void disconnect() {
+    _listener?.cancel();
+    _socket?.sink?.close(goingAway);
+    _socket = null;
+    _controller.add(false);
   }
 
   @override
   void dispose() {
-    _listener?.cancel();
-    _socket?.sink?.close(goingAway);
-    _socket = null;
+    disconnect();
+    _controller.close();
     super.dispose();
   }
 
@@ -44,17 +69,43 @@ mixin TCPSocketConnector on LifeCycle, Parser implements Connector {
   Socket _socket;
   StreamSubscription _listener;
 
+  StreamController _controller;
+
+  Stream get statusStream => _controller?.stream;
+
+  @override
+  void init() {
+    _controller = StreamController.broadcast();
+    super.init();
+  }
+
   @override
   Future<void> connect(String host, int port, {dynamic args}) async {
-    _socket = await Socket.connect(host, port);
-    _listener = _socket.listen(handleMessage);
+    try {
+      _socket = await Socket.connect(host, port);
+      _listener = _socket.listen(handleMessage, onError: (e) {
+        _controller.addError(e);
+        disconnect();
+      });
+      _controller.add(true);
+    } catch (e) {
+      _controller.addError(e);
+      _controller.add(false);
+    }
+  }
+
+  @override
+  void disconnect() {
+    _listener?.cancel();
+    _socket?.close();
+    _socket = null;
+    _controller.add(false);
   }
 
   @override
   void dispose() {
-    _listener?.cancel();
-    _socket?.close();
-    _socket = null;
+    disconnect();
+    _controller.close();
     super.dispose();
   }